requirements.txt
streamlit==1.25.0
streamlit-chat==0.1.1
google-ai-generativelanguage==0.2.0
google-api-core==2.11.1
google-auth==2.22.0
google-generativeai==0.1.0
googleapis-common-protos==1.59.1
grpcio==1.56.2
grpcio-status==1.56.2
